Implemented manual control and home pages for gui

+++ python_execute.py
@@ -3,6 +3,7 @@ import os
 import sys
 import signal
 import asyncio
+import psutil
 # from opentrons import api
 # from coordinator import *
 
@@ -15,7 +16,7 @@ WINDOWS_OS = 'nt'
 
 class Py_Execute:
     def __init__(self):
-        self.filename: str = None
+        self.filename: str = 'no_name.py'
         self.calibration_file_name: str = 'calibration.json'
         self.syringe_model: str = 'HAMILTON_175'
         self.p = None
@@ -76,22 +77,34 @@ class Py_Execute:
 
     def pause_execution(self):
         print("Pausing execution")
-        os.kill(self.p.pid, signal.SIGSTOP)
+        if os.name == LINUX_OS:
+            os.kill(self.p.pid, signal.SIGSTOP)
+        elif os.name == WINDOWS_OS:
+            self.p.kill()
+        
 
     def continue_execution(self):
         print("Continuing execution")
-        os.kill(self.p.pid, signal.SIGCONT)
+        if os.name == LINUX_OS:
+            os.kill(self.p.pid, signal.SIGCONT)
+        elif os.name == WINDOWS_OS:
+            psProcess = psutil.Process(self.p.pid)
+            psProcess.resume()
+            print("Trying to find a way to continue an execution on windows.")
 
     def stop_execution(self):
         print("Terminate protocol")
-        subprocess.Popen.terminate(self=self.p)
-
+        if os.name == LINUX_OS:
+            subprocess.Popen.terminate(self=self.p)
+        elif os.name == WINDOWS_OS:
+            pass
+            
     def info_from_protocol(self) -> list:
         labware_calibration_file_name = "None set"
         author = "None set"
         description = "None set"
         try:
-            if self.filename == None or self.filename == "- Select a Protocol -":
+            if self.filename == 'no_name.py' or self.filename == "- Select a Protocol -":
                 print("No file has been selected")
                 contents = "None set"
                 labware_calibration_file_name = "None set"
